Changes for filepath permission and other small fixes.
This commit changes the behavior of the filepath walk for metric baseline files to allow for files or folders that are not readable. It will now log those at the WARN level but continue. This is being introduced to address some OSX permission issues in the temp directory that were causing unit tests to fail locally. While unlikely to be encountered in production, it is likely preferrable behavrior in that environment, as well, and will leave logs to help direct debugging efforts if it is in error.

Additional changes in this commit are to remove dep as a require executable in the Makefile and to introduce some better log messages in the kubernetes_test.go file, again, around permissions.