Allow templating of additional labels on resources
This commit introduces two optional values - additionalPodLabels and additionalLabels. The former will apply the supplied labels on to the templated pod only, whereas the latter will apply them to all resources - including the pod.

These labels are applied in addition to the existing labels that the chart supplies. This therefore allows users to supply their own additional labels without affecting the essential labels supplied by the chart.

This is useful for situations where a user needs to supply additional labels for audit/attribution purposes... including attributing cost within Cloudability itself!

The deployment template was changed to include all labels on the pod, rather than just the selectorLabels. This does not impact the Deployment's selector, as it only looks for a subset of labels - not complete equality. As a result, even if a user does not supply any additional labels, this change will result in their pod gaining the app.kubernetes.io/version and app.kubernetes.io/managed-by labels that are already present on all other resources.